создайkf конструктор Picture
