+++ src/js/pictures.js
@@ -1,12 +1,13 @@
 'use strict';
 
-var getPictureTemplate = require('../js/getPictureTemplate.js');
+var Picture = require('../js/picture.js');
 var Gallery = require('../js/gallery.js');
 
 var renderPictures = function(pictures) {
   var container = document.querySelector('.pictures');
-  pictures.forEach(function(elem, pictureIndex) {
-    container.appendChild(getPictureTemplate(elem, pictureIndex++));
+  pictures.forEach(function(data, pictureIndex) {
+    var picture = new Picture(data, pictureIndex);
+    container.appendChild(picture.element);
   });
   Gallery.setPictures(pictures);
 };
